requirements.txt
Markdown
psycopg2-binary
pycrypto
Pygments
pillow
matplotlib
redis
Django
django-autoslug
django-celery
django-crispy-forms
django-extensions
django-filters
django-gravatar2
django-polymorphic
django-pygments
django-tables2
django-redis
hiredis
